Fix typo in kvdb iterator. Update format strings for compatibility

+++ src/fdb.c
@@ -62,6 +62,6 @@ void _fdb_init_finish(fdb_db_t db, fdb_err_t result)
         }
     } else {
         FDB_INFO("Error: %s (%s) is initialize fail (%d).\n", db->type == FDB_DB_TYPE_KV ? "KVDB" : "TSDB",
-                db->name, result);
+                db->name, (int)result);
     }
 }
